# backend/docs/api-spec.yaml
openapi: 3.0.0
info:
  title: Felparkering API
  description: Rest API för att hantera felparkerade fordon
  contact: {}
  version: '1.0'
servers:
  - url: http://localhost:8080
    variables: {}
paths:
...
  /reports:
    post:
      tags:
        - Reports
      summary: Create new report
      description: |-
        ### POST /reports

        Skapar en felparkeringsrapport.

        #### Behörighet:

        - `ADMIN`: Ja
            
        - `ATTENDANT`: Nej (403)
            
        - `CUSTOMER`: Ja
      operationId: Createnewreport
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                location:
                  type: string
                  example: Södra vägen 1, Stad
                licensePlate:
                  type: string
                  example: ABC123
                category:
                  type: string
                  example: Blockerar utfart
      responses:
        '201':
          description: Created
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ReportCreatedResponse'
                  - example:
                      message: Report created successfully
                      createdOn: '2025-06-17T14:22:00'
              example:
                message: Report created successfully
                createdOn: '2025-06-17T14:22:00'
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorMissingCredentials'
                  - example:
                      error: Missing credentials
              example:
                error: Missing credentials
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorUnauthorized'
                  - example:
                      error: Missing or invalid token
              example:
                error: Missing or invalid token
        '403':
          description: Forbidden
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorAccessDenied'
                  - example:
                      error: You do not have permission to create a report
              example:
                error: You do not have permission to create a report
      deprecated: false
      security:
        - bearer: []
    get:
      tags:
        - Reports
      summary: Get all reports
      description: |-
        ### GET /reports

        Hämtar en lista med alla felparkeringsrapporter.

        #### Behörighet:

        Roll:

        - `ADMIN`: Ser alla rapporter
            
        - `ATTENDANT`: Ser endast rapporter för sin egen vaktgrupp
            
        - `CUSTOMER`: Ser endast rapporter som de har skapat
      operationId: Getallreports
      parameters: []
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Report'
                description: ''
                example:
                  - id: 12
                    location: Södra vägen 1, Stad
                    licensePlate: ABC123
                    category: Blockerar utfart
                    createdOn: '2025-06-17T14:22:00'
                    updatedOn: '2025-06-17T14:22:00'
                    status: NEW
                    attendantGroup:
                      id: 1
                      namn: Stockholm Södra
                    assignedTo: 7
                  - id: 13
                    location: Norra vägen 3, Stad
                    licensePlate: ABC123
                    category: Saknar tillstånd
                    createdOn: '2025-06-17T15:17:00'
                    updatedOn: '2025-06-17T15:17:00'
                    status: NEW
                    attendantGroup:
                      id: 1
                      namn: Stockholm Södra
                    assignedTo: 7
              example:
                - id: 12
                  location: Södra vägen 1, Stad
                  licensePlate: ABC123
                  category: Blockerar utfart
                  createdOn: '2025-06-17T14:22:00'
                  updatedOn: '2025-06-17T14:22:00'
                  status: NEW
                  attendantGroup:
                    id: 1
                    namn: Stockholm Södra
                  assignedTo: 7
                - id: 13
                  location: Norra vägen 3, Stad
                  licensePlate: ABC123
                  category: Saknar tillstånd
                  createdOn: '2025-06-17T15:17:00'
                  updatedOn: '2025-06-17T15:17:00'
                  status: NEW
                  attendantGroup:
                    id: 1
                    namn: Stockholm Södra
                  assignedTo: 7
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorUnauthorized'
                  - example:
                      error: Missing or invalid token
              example:
                error: Missing or invalid token
        '403':
          description: Forbidden
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorAccessDenied'
                  - example:
                      error: You do not have permission to get all reports
              example:
                error: You do not have permission to get all reports
      deprecated: false
      security:
        - bearer: []
  /reports/{id}:
    get:
      tags:
        - Reports
      summary: Get a report
      description: |-
        ### GET /reports/{id}

        Hämtar detaljerad information om en specifik felparkeringsrapport.

        #### Behörighet:

        - `ADMIN`: Har tillgång till alla rapporter
            
        - `ATTENDANT`: Har tillgång om rapporten tillhör samma vaktgrupp
            
        - `CUSTOMER`: Har tillgång om rapporten är skapad av användaren
            

        #### Parameter

        - `id` _(Long)_ – rapportens ID
      operationId: Getareport
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          style: simple
          schema:
            type: string
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Report'
                  - example:
                      id: 12
                      location: Södra vägen 1, Stad
                      licensePlate: ABC123
                      category: Blockerar utfart
                      createdOn: '2025-06-17T14:22:00'
                      updatedOn: '2025-06-17T14:22:00'
                      status: NEW
                      attendantGroup:
                        id: 1
                        namn: Stockholm Södra
                      assignedTo: 7
              example:
                id: 12
                location: Södra vägen 1, Stad
                licensePlate: ABC123
                category: Blockerar utfart
                createdOn: '2025-06-17T14:22:00'
                updatedOn: '2025-06-17T14:22:00'
                status: NEW
                attendantGroup:
                  id: 1
                  namn: Stockholm Södra
                assignedTo: 7
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorUnauthorized'
                  - example:
                      error: Missing or invalid token
              example:
                error: Missing or invalid token
        '403':
          description: Forbidden
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorAccessDenied'
                  - example:
                      error: You do not have permission to get a report
              example:
                error: You do not have permission to get a report
        '404':
          description: Not Found
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorNotFound'
                  - example:
                      error: Report not found
              example:
                error: Report not found
      deprecated: false
      security:
        - bearer: []
    put:
      tags:
        - Reports
      summary: Update report status
      description: |-
        ### PUT/reports/{id}

        Uppdaterar informationen för en specifik felanmälan, t.ex. ändrar dess status.

        #### Behörighet:

        - `ADMIN`: kan uppdatera alla fält
            
        - `ATTENDANT`: kan ändra status om rapporten tillhör deras vaktgrupp
            
        - `CUSTOMER`: ej tillåtet (403)
            

        #### Parameter

        - `id` _(Long)_ – rapportens ID
      operationId: Updatereportstatus
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          style: simple
          schema:
            type: string
      responses:
        '403':
          description: Forbidden
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorAccessDenied'
                  - example:
                      error: You do not have permission to update a report
              example:
                error: You do not have permission to update a report
        '200':
          description: Ok
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Report'
                  - example:
                      id: 12
                      location: Södra vägen 1, Stad
                      licensePlate: ABC123
                      category: Blockerar utfart
                      createdOn: '2025-06-17T14:22:00'
                      updatedOn: '2025-06-17T14:22:00'
                      status: NEW
                      attendantGroup:
                        id: 1
                        namn: Stockholm Södra
                      assignedTo: 7
              example:
                id: 12
                location: Södra vägen 1, Stad
                licensePlate: ABC123
                category: Blockerar utfart
                createdOn: '2025-06-17T14:22:00'
                updatedOn: '2025-06-17T14:22:00'
                status: NEW
                attendantGroup:
                  id: 1
                  namn: Stockholm Södra
                assignedTo: 7
      deprecated: false
      security:
        - bearer: []
...
components:
  schemas:
...
    ErrorMissingCredentials:
      title: ErrorMissingCredentials
      required:
        - error
      type: object
      properties:
        error:
          type: string
      example:
        error: Missing credentials
...
    ErrorNotFound:
      title: ErrorNotFound
      required:
        - error
      type: object
      properties:
        error:
          type: string
      example:
        error: User not found
...
    ReportCreatedResponse:
      title: ReportCreatedResponse
      required:
        - message
        - createdOn
      type: object
      properties:
        message:
          type: string
        createdOn:
          type: string
      example:
        message: Report created successfully
        createdOn: '2025-06-17T14:22:00'
    AttendantGroup:
      title: AttendantGroup
      required:
        - id
        - namn
      type: object
      properties:
        id:
          type: integer
          format: int32
        namn:
          type: string
      example:
        id: 1
        namn: Stockholm Södra
    ErrorUnauthorized:
      title: ErrorUnauthorized
      required:
        - error
      type: object
      properties:
        error:
          type: string
      example:
        error: Missing or invalid token
    ErrorAccessDenied:
      title: ErrorAccessDenied
      required:
        - error
      type: object
      properties:
        error:
          type: string
      example:
        error: You do not have permission to create a report
    Report:
      title: Report
      required:
        - id
        - location
        - licensePlate
        - category
        - createdOn
        - updatedOn
        - status
        - attendantGroup
        - assignedTo
      type: object
      properties:
        id:
          type: integer
          format: int32
        location:
          type: string
        licensePlate:
          type: string
        category:
          type: string
        createdOn:
          type: string
        updatedOn:
          type: string
        status:
          type: string
        attendantGroup:
          $ref: '#/components/schemas/AttendantGroup'
        assignedTo:
          $ref: '#/components/schemas/User'
      example:
        id: 12
        location: Södra vägen 1, Stad
        licensePlate: ABC123
        category: Blockerar utfart
        createdOn: '2025-06-17T14:22:00'
        updatedOn: '2025-06-17T14:22:00'
        status: NEW
        attendantGroup:
          id: 1
          namn: Stockholm Södra
        assignedTo: 7
...
    User:
      title: User
      required:
        - id
        - email
        - role
        - attendantGroup
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
        role:
          type: string
        attendantGroup:
          oneOf:
            - allOf:
                - $ref: '#/components/schemas/AttendantGroup'
                - example:
                    id: 1
                    namn: Stockholm Södra
            - nullable: true
      example:
          id: 7
          email: Stockholm Norra
          role: ATTENDANT
          attendantGroup:
            id: 1
            namn: Stockholm Södra
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...