# backend/src/main/java/se/voizter/felparkering/api/model/User.java
package se.voizter.felparkering.api.model;

import jakarta.persistence.*;
import se.voizter.felparkering.api.enums.Role;

@Entity
@Table(name = "users")
public class User {
    
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Column(nullable = false, unique = true)
    private String email;

    @Column(nullable = false)
    private String password;

    @Enumerated(EnumType.STRING)
    @Column(nullable = false)
    private Role role;

    @ManyToOne
    @JoinColumn(name = "attendant_group_id")
    private AttendantGroup attendantGroup;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    public Role getRole() {
        return role;
    }

    public void setRole(Role role) {
        this.role = role;
    }

    public AttendantGroup getAttendantGroup() {
        return attendantGroup;
    }

    public void setAttendantGroup(AttendantGroup attendantGroup) {
        this.attendantGroup = attendantGroup;
    }
}